Moved marshalling code out of model
