+++ pkg/loghttp/legacy/query.go
@@ -2,8 +2,6 @@ package loghttp
 
 import (
 	"time"
-
-	"github.com/grafana/loki/pkg/logproto"
 )
 
 // QueryResponse represents the http json response to a label query
@@ -21,26 +19,4 @@ type Stream struct {
 type Entry struct {
 	Timestamp time.Time `json:"ts"`
 	Line      string    `json:"line"`
-}
-
-// NewStream constructs a Stream from a logproto.Stream
-func NewStream(s *logproto.Stream) *Stream {
-	ret := &Stream{
-		Labels:  s.Labels,
-		Entries: make([]Entry, len(s.Entries)),
-	}
-
-	for i, e := range s.Entries {
-		ret.Entries[i] = NewEntry(e)
-	}
-
-	return ret
-}
-
-// NewEntry constructs an Entry from a logproto.Entry
-func NewEntry(e logproto.Entry) Entry {
-	return Entry{
-		Timestamp: e.Timestamp,
-		Line:      e.Line,
-	}
 }
