+++ pkg/loghttp/query.go
@@ -6,12 +6,7 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/prometheus/prometheus/promql"
-
-	"github.com/grafana/loki/pkg/logproto"
-	"github.com/grafana/loki/pkg/logql"
 	"github.com/prometheus/common/model"
-	"github.com/prometheus/prometheus/pkg/labels"
 )
 
 // QueryStatus holds the status of a query
@@ -107,106 +102,3 @@ type Vector []model.Sample
 
 // Matrix is a slice of SampleStreams
 type Matrix []model.SampleStream
-
-// NewStreams constructs a Streams from a logql.Streams
-func NewStreams(s logql.Streams) (Streams, error) {
-	var err error
-	ret := make([]Stream, len(s))
-
-	for i, stream := range s {
-		ret[i], err = NewStream(stream)
-
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return ret, nil
-}
-
-// NewStream constructs a Stream from a logproto.Stream
-func NewStream(s *logproto.Stream) (Stream, error) {
-	labels, err := NewLabelSet(s.Labels)
-	if err != nil {
-		return Stream{}, err
-	}
-
-	ret := Stream{
-		Labels:  labels,
-		Entries: make([]Entry, len(s.Entries)),
-	}
-
-	for i, e := range s.Entries {
-		ret.Entries[i] = NewEntry(e)
-	}
-
-	return ret, nil
-}
-
-// NewEntry constructs an Entry from a logproto.Entry
-func NewEntry(e logproto.Entry) Entry {
-	return Entry{
-		Timestamp: e.Timestamp,
-		Line:      e.Line,
-	}
-}
-
-// NewVector constructs a Vector from a promql.Vector
-func NewVector(v promql.Vector) Vector {
-	ret := make([]model.Sample, len(v))
-
-	for i, s := range v {
-		ret[i] = NewSample(s)
-	}
-
-	return ret
-}
-
-// NewSample constructs a model.Sample from a promql.Sample
-func NewSample(s promql.Sample) model.Sample {
-
-	ret := model.Sample{
-		Value:     model.SampleValue(s.V),
-		Timestamp: model.Time(s.T),
-		Metric:    NewMetric(s.Metric),
-	}
-
-	return ret
-}
-
-// NewMatrix constructs a Matrix from a promql.Matrix
-func NewMatrix(m promql.Matrix) Matrix {
-	ret := make([]model.SampleStream, len(m))
-
-	for i, s := range m {
-		ret[i] = NewSampleStream(s)
-	}
-
-	return ret
-}
-
-// NewSampleStream constructs a model.SampleStream from a promql.Series
-func NewSampleStream(s promql.Series) model.SampleStream {
-	ret := model.SampleStream{
-		Metric: NewMetric(s.Metric),
-		Values: make([]model.SamplePair, len(s.Points)),
-	}
-
-	for i, p := range s.Points {
-		ret.Values[i].Timestamp = model.Time(p.T)
-		ret.Values[i].Value = model.SampleValue(p.V)
-	}
-
-	return ret
-}
-
-// NewMetric constructs a labels.Labels from a model.Metric
-func NewMetric(l labels.Labels) model.Metric {
-	ret := make(map[model.LabelName]model.LabelValue)
-
-	for _, label := range l {
-		ret[model.LabelName(label.Name)] = model.LabelValue(label.Value)
-	}
-
-	return ret
-}
