+++ pkg/logql/marshal/marshal.go
@@ -1,4 +1,4 @@
-package loghttp
+package marshal
 
 import (
 	"encoding/json"
@@ -6,6 +6,7 @@ import (
 	"io"
 
 	"github.com/gorilla/websocket"
+	"github.com/grafana/loki/pkg/loghttp"
 	legacy "github.com/grafana/loki/pkg/loghttp/legacy"
 
 	"github.com/grafana/loki/pkg/logql"
@@ -18,12 +19,12 @@ import (
 func WriteQueryResponseJSON(v promql.Value, w io.Writer) error {
 
 	var err error
-	var value ResultValue
-	var resType ResultType
+	var value loghttp.ResultValue
+	var resType loghttp.ResultType
 
 	switch v.Type() {
-	case ResultTypeStream:
-		resType = ResultTypeStream
+	case loghttp.ResultTypeStream:
+		resType = loghttp.ResultTypeStream
 		s, ok := v.(logql.Streams)
 
 		if !ok {
@@ -35,8 +36,8 @@ func WriteQueryResponseJSON(v promql.Value, w io.Writer) error {
 		if err != nil {
 			return err
 		}
-	case ResultTypeVector:
-		resType = ResultTypeVector
+	case loghttp.ResultTypeVector:
+		resType = loghttp.ResultTypeVector
 		vector, ok := v.(promql.Vector)
 
 		if !ok {
@@ -44,8 +45,8 @@ func WriteQueryResponseJSON(v promql.Value, w io.Writer) error {
 		}
 
 		value = NewVector(vector)
-	case ResultTypeMatrix:
-		resType = ResultTypeMatrix
+	case loghttp.ResultTypeMatrix:
+		resType = loghttp.ResultTypeMatrix
 		m, ok := v.(promql.Matrix)
 
 		if !ok {
@@ -57,9 +58,9 @@ func WriteQueryResponseJSON(v promql.Value, w io.Writer) error {
 		return fmt.Errorf("v1 endpoints do not support type %s", v.Type())
 	}
 
-	j := QueryResponse{
+	j := loghttp.QueryResponse{
 		Status: "success",
-		Data: QueryResponseData{
+		Data: loghttp.QueryResponseData{
 			ResultType: resType,
 			Result:     value,
 		},
